Update TimerFunctions.js
Commented in a few notes.  Adjusted the way Variables are declared.  You should use access modifiers to improve on OOP.  It can be annoying at first, but it is strongly recommended.  Line: 13.  Not sure if this is an issue in JS or not, but in AS you needs to set the values to that type of value or they are converted right away.  Very annoying; Not sure if it's an issue in JS though.